requirements.txt
fastapi
uvicorn
sqlalchemy
databases
pydantic
python-jose
passlib[bcrypt]
hydra-core
omegaconf
-e .